Submit PyPI exercise
