+++ src/chinhalidiffusion2D.egg-info/requires.txt
@@ -1,0 +1,2 @@
+NumPy
+Matplotlib
